+++ doc/release_notes.rst
@@ -1,6 +1,8 @@
 Release notes
 =============
 
+* Module ximpol.core.fitsio added (issue #49).
+
 
 *ximpol (0.15.2) - Mon, 25 Jan 2016 18:04:33 -0800*